Backend: Backport the webserver from flamescope
Use the flamescope webserver and the commit id to base on is:
6abdc8 task: upgrade dependencies

To start the webserver:
python3 ./run.py

Before that you would like to prepare the frontend in /build folder. The
supported frontend source code will be commited later. And it can be
built by "npm run build" and you would like to increase the runtime
memory size by: "export NODE_OPTIONS=--max_old_space_size=4096" to avoid
the OOM error in the build time.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>