Frontend: Integrate the frontend to link to webserver
The front end TableRow needs to connect to the webserver of flamescope.
Here, we just have an example how to do that. In the future, the backend
side needs to adjust to scan the available projects under the example
folder and send to the frontend to show on the table rows. And the Link
to address needs to be modified to route to dashboard.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>