Frontend/Backend: Support the project list scan
Modify the code to support the projects scanning and show up the
available projects when clicking on the "dashboard" sidebar. By default,
the projects is put in the "examples" folder. And the available projects
is in the form of folders. Only the folders will be showed up in the
available projects list. When clicking on the project name table row,
the page will be routed into the corresponding dashboard of the project.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>