Frontend/webpack: Add the configuration of webpack
To run the development server: "npm run dev". Then, the localhost:8080
will be routed to localhost:5000 for some specific
fetches(e.g. dashboard/profile) which are recorded in the
webpack.config.js. The development version can dynamically rebuild when
the .js files are changed. However, the rebuild is saved in memory and
main.js is not changed.

To build the production version: "npm run webpack". The main.js will be
generated to build/. And the index.html is generated by refering to the
template under src/templates/index.template.ejs.

The coding style/grammar can be checked by: "npm run lint". And before
that you may want to run "npm run eslintInit".

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>