Frontend: Remove the unnecesssary files in the build folder
Most of the files under build are unnecessary and need to be removed.
Only keep the index.html and main.js. The index.html is generated
automatically by "nvm run webpack" and it's referred to
src/templates/index.template.ejs. This is recorded in webpack.config.js.
The main.js is also generated and defined in webpack.config.js too.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>