Frontend/Line: Add the Zoomable Line Chart implementation and use the d3 V6 as default
The zoomable Line chart is referred to:
https://www.d3-graph-gallery.com/graph/line_brushZoom.html

And I added one more line for multiple line testing. The other line is
referred to the original data by shifting a bit of the Y axis value.

Frontend/Line: Use the d3 V6 as default

Use the d3v6 as the default version of d3. However, the Scatterv5 needs
the v5. So, add the d3v5 and import that into the Scatterv5 to not break
the scatterplot.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>