Frontend/Line: Modify the zoom-out behavior to use right click
If single left-click[1] is used as the way to change the corresponding
flamegraph/scatterplot/... chart adjusting method, the double left-click
zoom out will also receive the unnecessary event of single left-click.
And it's the default behavior of the browser. So, change the zoom-out to
right-click to avoid that. Also the pop-menu of the right-click needs to
be avoided by d3.event.preventDefault()[1].

Some one writing the code to distinguish that[2]. But it's redundant to
write code to avoid that. Just change the behavior is easier.

[1]. right-click on a d3.js element: how to prevent browser context menu to appear
https://stackoverflow.com/questions/24653471/right-click-on-a-d3-js-element-how-to-prevent-browser-context-menu-to-appear

[2]. Distinguishing click and double-click in D3.
http://bl.ocks.org/couchand/6394506

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>