Frontend/Line: Fix the wrong return coordinates when click
When use arrow function as the callback function, the "this" points to
the outside window component and is not the svg component itself. The
problem is that the returned coordinate will be the offset of the
windows component. So, it will make the reverse calculation of the date
inside the Line component wrong:

"const {date, value} = bisect(d3.pointer(event, this)[0]);"

After looking at the document[1]:

"d3.pointer(event[, target])
If target is not specified, it defaults to the source event’s
currentTarget property, if available. If the target is an SVG element,
the event’s coordinates are transformed using the inverse of the screen
coordinate transformation matrix. If the target is an HTML element, the
event’s coordinates are translated relative to the top-left corner of
the target’s bounding client rectangle. (As such, the coordinate system
can only be translated relative to the client coordinates. See also
GeometryUtils.) Otherwise, [event.pageX, event.pageY] is returned."

We need to use arrow function as the state of the flamegraph needs to be
changed when clicking on the line. Without the arrow function, the
callback function of the changeState from Memory.js cannot be changed:

"<Line changeState={this.changeState}/>"

So, we don't use "this" here to get the coordinates according to the svg
component:

"const {date, value} = bisect(d3.pointer(event)[0]);"

[1]. https://github.com/d3/d3-selection

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>