Profile: The default state should keep all the temporary files
When G_KEEP_TAR and G_KEEP_SVG are all false, all the temporary files
should be kept. So, '-d' is not a correct behavior as it keeps only the
.svg file. So, remove the '-d' and keep it as '' to meet the default
behavior.

On the other hand, the tristate is much easier to put all the
configuration into one option and don't need to use
G_KEEP_TAR/G_KEEP_SVG. In the future, this part should be improved.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>