Profile: Move the profile-related code to profile folder
As the frontend/backend are implemented to the system, the root folder
is becoming more complicated and hard to organize. Put the system
profiling-related code to profile folder to make source code more
organized and easy to manage.

However, the gen-flamegraph.sh is still kept in the root folder. The
intention is to make people to easily use the script to parse the
perf.data to flamegraph. Also, there are no relationship to other script
in local usage. The dependence only happes in the
/usr/lib/easy-flamegraph and the script is installed by Makefile. So,
the location inside the /usr/lib/easy-flamegraph is the same.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>