Profile: The package installation should be on-demand
The package installation should be on-demand instead of installing all.
Currently, some of the package is not necessary on general profiling.
For example, the "bcache-tools" package is not a popular package as most
of the scenarios doesn't have the bcache installed. The patch use the
configuration inside the easy-flamegraph.config as the principle to
determine if the required tools should be installed.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>