Profile: Support the {net,irq}-stat profiler
These two profilers are from the idea of Niv and the disussion between
Jay/Matthew. The useful information can be used to solve the networking
performance problem especially in the RPS/RSS packet distribution
problem with high CPU usage with ksoftirqd on a specific processor.

The detail flamegraph can be referred to:
112.1 host: CPU 3/35 has high CPU ksoftirqd
112.2 host: CPU 0/32 has high CPU ksoftirqd

https://people.canonical.com/~gavinguo/292844/

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>