Profile/cleansing: Add the mem-extract.py and move-svg.py
The mem-extract.py is used for extracting the free memory information
from meminfo, then, the data will transfer to the csv file. The
move-svg.py is used for transfering the .script callstack text to the
naming format without the seconds. In the future, the mem-extract needs
to extend to support all the parsing work of memory-related data.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>