Profile/subsystem/ftracer.py: Implement the Ftrace sampling
Currently, impelmenting the perf memory allocation sampling is easy.
However, the performance overhead of running perf in the production
system is high. Also, the customer needs to install the linux-tools for
perf to run in the production system. On the other hand, the default
Ftrace inside the kernel can help to capture the call graph and the
latency is included. Instead, the flamegraph can be generated from the
ftrace. The latency chart and further analysis can be done based on the
Ftrace log analysis.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>