Profile/conditions/io: Add the io_ftrace implementation
The current implementation with perf only profiles the io event of
callstack. However, we need to know the detail IO latency for a specific
request. By using ftrace, the latency can be calculated with
blk_mq_start_request/blk_account_io_completion clock difference. Also,
the stacktrace can be echoed into the trigger option. This will be light
weight and much easier to use in the production system compared to perf.

In the future, there should be at least multiple backend profiling
techniques supported. For example:

1). ftrace 2). bcc 3). perf

However, the trace_clock still needs to be fine-tuned to select a best
one as currently [perf] is used as the testing clock and it seems good.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>