Profile/conditions/mem: Add the implementation of mem_ftrace
The original implementation of memory profiling uses the perf as the
backend. However, the perf has serious performance overhead in the
production environment. And there are limitations to get the detail
information from the system. To avoid the dependence on the perf and
increase the flexibility of captures, add ftrace as the backend.

There are many ideas to further implement the data science skills such
as data visualization, projection on different features, and regression
prediction on the memory usage, etc.

The user space sampling in the original profiling needs to consider how
to implement in the Ftrace framework. The old mem script is renamed to
profile/conditions/mem/mem_old for future reference and can be deleted
unitl all the cases are reimplemented. The features include
tcmalloc/malloc/threshold/process sampling, etc.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>