gen-flamegrph.sh: Add the support with ftrace/trace-cmd parsing
The support to the ftrace/trace-cmd stackcollapse parsing is already in
my implemented stackcollapse-tracecmd.pl. Then, the gen-flamegraph.sh
needs to update to comply with both ftrace/perf format. There are still
some processes which are not complied with perf and ftrace. For example,
the per-cpu flamegraph implementation still needs to implement with
ftrace input. Maybe to cat the trace file under each CPU or it can be
separated from the input stack as there is the CPU number before each
stack output. Other than that, the code is also refactor to make it
readable.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>