gen-flamegraph.sh: Move the perf.data header extraction outside __generate_flamegraph
Currently, the perf.data header information extraction is inside the
__generate_flamegraph function. However, it will parse the perf.data
every time in the per-cpu flamegraph generating loop. However, it just
needs to parse one time as the system information is the same across
different CPUs and the performance impact is obvious when the header
extraction is enabled by default. On the other hand, the header check is
missed in the previous porting and needs to add back to make it work
again.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>