Profile/cleansing: Reimplement by argparse and add unifying the csv
The original arguments parsing is a bit trivial and cannot support many
error checking. Reimplement the arguments by argparse. Also, the csv was
generated separately by different parameters. Add the support to
generate an unifying one to make it easy for the small file case. The
original idea to generate separate ones are to eliminate the size and
not all the parameters are useful in generating the chart/graph.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>