Profile: Support meminfo/vmstat line chart
The time series data chart is important to diagnose the performance
problem in a specific of time. easy-flamegraph already have the
periodically sampling code in the /profile/sysinfo. The
profile/cleansing/mem-extract.py is used to generate the
{meminfo,vmstat}.csv.

The current implementation is still not user-friendly and needs to
improve in the following directions:

- The current implementation is to read the .csv
file in the current directory. It needs to be improved asap to import
from 1). a specific options: "gen-flamegraph.py --meminfo
~/Downloads/meminfo.csv" 2). a specified folder: "gen-flamegraph.py --folder
~/Downloads/easy-flamegraph"

- The current Y axis name is hard-coded.

- The group list inside the csv files need to be generalized into a
config instead of hard-coded.

- The existence of parameters needs to be checked. For example, the
nr_file_hugepages in /prov/vmstat doesn't exist in the old kernel and
will throw out the error in runtime when parsing the old kernel data.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>