Profile/cleansing: Support the tap links for the memory graph
The tap link when clicking on the dot of a specific date will open a new
tab for a specific svg file. The current implementation is to open the
memory allocation flamegraph related to the date. The future improvement
could be opening a summary for the moment. For example, the RSS/swap
proportion.

Also, there are some error checkings need to be implemented:
1). The absence of the svg files for some reasons(low memory) and the
tap link should be NULL.

2). plot_meminfo/plot_vmstat has many things in command. It will be good
to extract the common part and put in the common function to make it
simple and clear.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>