summary.py: Implement a wrapper to generate the dashboard
The current implementation is to parse the csv by the script under:
profile/cleansing/. For example, profile/cleansing/mem-extract.py.

After transfer the captured data into the csv files. Next, the dashboard
needs to be generated for each subsystem. The profile/dashboard.py does this.

For the people not familiar with the tool to understand how
to use the tool, and it's better to have a wrapper to generate the
dashboard with one command. So that's why summary.py exist.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>