profile/conditions/cpu: Implement the mpstat-P-ALL capture
To implement the dashboard for each processor CPU utilization, the
"mpstat -P ALL 1 1" is a prerequisite to capture the data for each
minute. However, the mpstat capture needs to combine with the idle rate
capturing to measure the CPU threshold in profile/conditions/cpu.  The
reason to do that is there will be the CPU utilization deviation in
capturing the data in a short different time within a minute. The
sysinfo capturing is in the same minute. However, the sysinfo capturing
is ahead of the CPU perf sampling(profile/entry). The time difference in
the sampling will make the dashboard not correct.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>