profile/entry: Fix the flock bug
The flock cannot work correctly when the G_PERF_PERIOD=100. Perf will
run for more than 1 minute. And it will overlay with the previous one if
the flock checking succeeds. Here, the implementation is only for the
extended low-level system profiling and is not applied to the sysinfo
capture. The inconsistent will make the counts between the records and
svg not working in the Dashboard. I will need to fix the summary.py to
make it work.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>