Dashboard: Support the CPU utilization chart
The CPU utilization chart (based on mpstat -P ALL 1 1) is handy when
debugging the high loading issue in the system. When clicking on the
specific date on the line, the corresponding flamegraph will be popped
out to a new tab.  Some points need to be aware of when using the tool:

sudo ./summary.py or ./summary.py -i ~/Downloads/easy-flamegraph/

Use sudo because the default output folder of summary.html is in
/var/log/easy-flamegraph.

1). The # of CPUs are limited to eight (8). When there are #CPUs > 8,
all system utilization will be generated instead of generating every
core line chart. This needs to improve by maybe using the heatmap of CPUs
when clicking on a specific time. Then, the heatmap can implement the
flamegraph link. The reason to not enable all CPUs chart is when there
are 128 CPUs, the chart cannot plot all the lines and will be
overwhelming in a small plot.

2). The origin of flamegraphs is from the input folder. It's in
/var/log/easy-flamegraph/{cpu,mem}.

3). When clicking on the dot, then the
file:///home/vin/Downloads/easy-flamegraph/cpu/NaN is popped out. The
problem is that the CPU_THRESHOLD is too high to trigger the performance
profiling. For example, CPU_THRESHOLD=50, then every loading smaller
than 50 will not be profiled. We need to improve by adding a horizontal
line for a specific loading with CPU_THRESHOLD to notify the limitation.

4). There are %irq, %gnice, %steal. The specific indicators can
implement with a different flamegraph than a typical CPU flamegraph.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>