profile/subsystem/memory.py: Add the memory index with v5.4
Capture the /proc/version and save in /var/log/easy-flamegraph/version.
The version is used for the kernel version checking and can add the
specific vmstat/meminfo parameters when the kernel version is > v5.4.
Currently, it can parse the v4.15 kernel parameter. The Xenial kernel
still needs to test to find the specific parameters for the v4.15 Bionic
kernel. The checking is for the Ubuntu LTS kernel.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>