profile/conditions/mem/: Capture the memory info before/after memory stack profiling
The memory status profiling can be compatible with the memory allocation
profiling. When the memory metrics are captured before/after the memory
allocation, the metrics can be compared to the memory Flamegraph to see
the source of the exctly amount of memory usage.

For example, the line chart can be identified where is the high
anonymous allocation. Then, click on the chart to look for the anonymous
memory allocation in the Flamegraph.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>