Remove the webserver implementation of Flamegraph
Easy-flamegraph has the webserver implementation, which has the d3
Flamegraph frontend and the flask backend. However, it's a heavy loading
to maintain the architecture. Currently, there is a Bokeh dashboard
implementation to demonstrate the data in an acceptable format. It's
time to remove the webserver implementation and back it up into the
webserver branch for future reference. Maybe this will be reopened if we
figure out a way to do that.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>