Support the thread pool agents model
The current shell script sampling using the crontab is not scalable when
there are multiple data agents to run. The prototype is to build a
thread pool framework. The current default thread is the number of CPUs
on the system. There are three essential components:

1). The thread pool
Goroutine implements the thread. And the number of the Goroutines is
tasks on the queue. When there is a new task, the Goroutines are
contending to get out of the tasks for further processing.

2). The task queue
The task queue is implemented in the task interface, which can accept
any data agent callback function implementations. The task is wrapped by
the data agent from different sources (CPU, memory, IO, ...).  For
example, if the perf record needs to run every minute, the work can be
wrapped in the task callback and inserted in the task queue for further
processing by Goroutines.

3). The data agent
The data agents are the different sources of data capturing. That
includes many other domains. For example, the perf record to capture the
CPU utilization cases. And the data agent needs to be scheduled to run
on the system for different periods.

To run the testing program:
$ cd profile
$ go run main.go
waiting
goroutine 0 in loop
test workfunction!!
goroutine 1 in loop
test workfunction!!
goroutine 4 in loop
test workfunction!!
goroutine 2 in loop
test workfunction!!
goroutine 1 in loop
test workfunction!!
goroutine 4 in loop
test workfunction!!
goroutine 2 in loop
test workfunction!!
goroutine 0 in loop
test workfunction!!
^Csignal: interrupt

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>